Version 5.2

+++ pwdmkr.py
@@ -1,4 +1,5 @@
-#pwdmkr v5.1 by maxrt101
+#pwdmkr v5.2 by maxrt101
+import os
 import sys
 sys.dont_write_bytecode = True
 import random
@@ -10,6 +11,7 @@ from functools import reduce
 parser = argparse.ArgumentParser(description='''Python password maker. Example: ./pwdmkr.py  -l 16  -m b  -d -  -dl 4 ''')
 parser.add_argument('-v', action='store_true', help='Display version and exit')
 parser.add_argument('-s', action='store_true', help='Save password to file')
+parser.add_argument('-fs', action='store_true', help='Force save password to file')
 parser.add_argument('-f', action='store', dest='file', help='Destination file to save', default='password.txt')
 parser.add_argument('-l', action='store', dest='length', help='Length of password', default=16)
 parser.add_argument('-m', action='store', dest='mode', help='Mode: letters only(l), numbers only(n) or both(b)', default='b')
@@ -19,7 +21,7 @@ parser.add_argument('-dl', action='store', dest='delimiter_len', help='Length be
 
 args = parser.parse_args()
 
-version = '5.1'
+version = '5.2'
 
 config = {
 	"len": 16,
@@ -52,10 +54,17 @@ def main():
 	p2 = reduce((lambda x, y: x + config["delimiter"] + y if ((len(x) + 1) % (config["delimiter_len"] + 1) == 0)  else x + y), list(p))
 	print(p2)
 
-	if args.s == True:
-		pwdfile = open(args.file, 'w')
-		pwdfile.write(p2)
-		pwdfile.close()
+	if args.s == True and args.fs == True:
+		print('''ERROR: Arguments -s and -fs can't be used together''')
+	else:	
+		if args.s == True or args.fs == True:
+			if os.path.isfile(args.file) == True and args.s == True:
+				print('ERROR: File {} already exists'.format(args.file))
+			else:
+				pwdfile = open(args.file, 'w')
+				pwdfile.write(p2)
+				pwdfile.close()
+
 
 
 if args.v == True:
